Add retry stack to vm. (#9)
This enables backtracking to work properly. Includes an additional test that permutes all possible inputs for the functions test. Adds the jealous test, and a new nested rule test.

The retry field in a choice tracks the goal (query) that lead to those choices. Then, when we backtrack we use the retry field to reconstruct the rest of the query after trying the next set of alternatives.

While currently query is the only instruction that yields choices, we could have others. For example EXTERNAL_LOOKUP could return multiple results, which requires multiple choice points an a retry.

Co-authored-by: Alex Plotnick <alex@osohq.com>