Update Polar#register_class to take a proc; not a block
This allows Oso#register_class to transform any block it receives into a
Proc and pass that along to Polar#register_class. It's the best way I
could find to accept a block in one method and pass it along to another
function that wants to accept a block-like object (a closure).