Descending into operations doesn't seem to infinitely recurse
in the same way descending into partials would, and it seems necessary
to capture the pre-expression bindings of variables that get hoovered up
into an expression. For example, if `x` is bound to `1` and we evaluate
the expression `x in y` where `y` is already an expression (f.k.a. a
partial), before we rebind `x` to the newly constrained expression we
need to ensure we deref `x` to `1` so that `x` and `y` both point to the
expression `1 in y`.