Fix exampleGet shortcode panic
Once the context changes via `{{- with $.Page.Resources.GetMatch
"data/data.md" -}}`, there's no more `.Position` attribute on the
current context (a resource.Resource struct instead of whatever
shortcode struct we started with), so Hugo would fail to render the
shortcode and panic. To fix it, we store .Position off in a variable
prior to spelunking down the `{{ with }}` tree.