Only construct rule_types for exercised rules
Previously we used the has_relations attribute in resource blocks to
generate the list of expected has_relation rule types based on the
unique related type pairs.

This could lead to false positive warnings if users mapped out their
model hierarchy with complete relationships ahead of leveraging them for
authorization rules.

We not limit the creation of has_relation rule types to only those
resource types which are explicitly used in shorthand rules.