Add "required" rule type generation
Extend RuleType with a `required` boolean which is used to denote that
the knowledge base *must* contain a rule which matches the rule type.

These "required" rule types are only created for resource relationships
which are actively consumed in shorthand rules.

Traverse the `declarations` for all resource types to create non-required
Rule Types for the described relations.