Add RuleSet#required_rule_types
Create a public method required_rule_types which returns the set of rule types
to traverse for validation obviating the need to expose the inner HashMap
directly.

Add comments to the create_resource_specific_rule_types code explaining the
new Rule struct "required" field and the scenarios in which it is set to `true`.