back out Actor||Resource rule type specializer & update lang suites
remove the special handling of the Actor and Resource symbols and
disallow the use of `Resource` as a specializer in rule types. This
means that rules of the form `has_role(user: User, role: String,
resource: Resource)` will have to be broken out into their
resource-specific definitions.

update the language integration test suites to stop using `Resource` as
a specializer