Support multiple resource blocks for the same resource (#1302)
Oso now allows multiple resource blocks to be declared for the same resource. The declarations from all resource blocks for a given type are merged together before policy evaluation. This permits rules in one block to reference declarations in another and for resource blocks to be composed over multiple files.