*Actually* verify that we got a "Done" event
Previously we were checking if looking up the "Done" property on the
received JS object resulted in an error, but really we care that it
doesn't error *and* returns an expected value. Looking up a non-existent
property on a JS object returns undefined. So in the new tests we added
in this PR (#1447) we were receiving "Result" events and then looking up
the "Done" property and receiving undefined w/o erroring.