{Python,Ruby,JS}: Add `class_id` information to ExternalInstance values; use for in-VM IsA checks (#1472)
Create `Constants` type which wraps two HashMaps describing Symbol ->
Term and ID -> Symbol relationships. This second ID -> Symbol index is
populated by reading the optional `class_id` information in
`ExternalInstance` and will be used to implement in-core IsA checks
where present.

Extend `vm#isa` to attempt an in-core check when the `ExternalInstance` value
we're processing has the necessary `class_id` information declared. At time of
commit Python is the only language which is capabale of passing this information
to the core.

* {JS,Python,Ruby} Format ExternalInstance values with user-provided class names if present  (#1488)

Co-authored-by: Gabe Jackson <17556281+gj@users.noreply.github.com>