Use a matrix for library build in release workflow
This is intended to make it easier to see what targets are currently
being built, and to simplify adding others in future.

It additionally introduces more parallelism, since each target will
build in parallel vs. each OS before (and atm each OS has two builds).
On the down side, it's not currently possible to depend on specific
matrix job instances, so some jobs have to wait for additional builds.
In theory, the parallelism should cancel this out, for a net win.

It's possible the naming conventions could be standardised such that the
`shared`/`static` fields could just be `boolean`, but this way no
changes are needed in the other jobs. Something to consider in future.