refactor: Clarify error when right_path exists
Fixes #1552

This commit catches the assertion error when verifying the `right_path`
does not exist when a path has been identified on the left side of the
policy. This refactor was made to better clarify the cause of the error
for developers who may not be familiar with the technical limitations of
the current SQLAlchemy integration.

In my case, I had assumed that the use of foreign keys and paths on both
sides would be more performant, and thus the best choice for a policy,
when writing the authorization rules. The resulting error didn't include
any context about what was wrong with the policy I had written. This
commit aims to address this by providing a best effort guess to provide
a starting point for developers to debug the policy issue themselves.