Clarify where there is no allowed action
Currently, when providing Oso to teams I work with, this error message causes a lot of confusion because they interpret it to mean that there is no allowed action for the entity in the policy. By clarifying where there is no allowed action (checked_permissions), this will alleviate this source of confusion and ensure that it's clear to the developer that they need to update checked_permissions in the application rather than in the policy.