Simplifies conditional logic using De Morgan's laws (#1586)
* Simplifies conditional logic using De Morgan's laws

X-Ref: https://github.com/osohq/oso/pull/1557

When reading logical statements it is important for them to be as simple as possible. This proposal applies De Morgan's laws to simplify code by removing double negatives or pushing negations deeper into statements. In this case, we're ensuring that the comparison against `None` is pythonically called (`is not None`)

Co-authored-by: Sam Scott <sam@osohq.com>