Replace mutable default argument with `None` (#1583)
A common gotcha in Python involves default argument values which are mutable. These are evaluated only once at function definition time, so only one list, set or dictionary instance will be created. This means that if this list is mutated in one call to a function, those changes will show up in subsequent calls of that function. This is usually unintended behaviour, though it can be useful in limited circumstances for writing caches.

This updates `load_files` to avoid the mutable default argument of `[]`. While the code is not currently mutating the list, this is a high-risk piece of code that will be a gotcha should the team choose to modify the code at a later date to mutate the argument.